Introduced DoubleLedger and testConsensusRange for ledger testing
The DoubleLedger does a lot of double checking on ledgers produced by
rerunning transactions after creating blocks, deserializing the txns
from the block and validating the results.

testConsensusRange makes it easy for ledger tests to test behavior in
many protocol levels. A particular problem it handles is that test can
be written to run from some consensus version and in every subsequent
version. So if we change behavior in a new version, we'll find out.