Add an unlimited global storage mechanism - Boxes.
At least three four left to consider, as well as extensive testing.

1. It has no LRU cache for the reads and writes.  I expect sqlite to
do a fine job caching kv pairs, and there's not savings from skipping
msgpack serialization. But sqlite has a page cache, not a row cache,
so maybe it's worth it.

2. There is no loop around the sql read with round number test.

3. There's is no locking of au.accountsMu.  This is clearly wrong,
need to learn more about the rules that govern when the other lookup
routines are _not_ taking lock.

4. Need to think through limits on total boxes read/written, and how
boxes can be "named" for access by new apps.